Allow users to add media content in the root home directory
The 'home' plug allows the snap to access the home directory of the user
executing the snap. Since the PMS snap runs as a daemon, it is executed
by the root user. Thus, with the 'home' plug enabled the snap is able to
access the home directory of the root user.